requirements.txt


langgraph>=1.0.3
langchain>=1.1.3
langchain-openai>=1.1.3
graphrag>=2.5.0
neo4j>=6.0.3
streamlit>=1.52.1
easyocr>=1.7.2
openai>=2.11.0
python-dotenv>=1.0.1
pydantic>=2.11.3
pyvis>=0.3.2
sentence-transformers>=3.0.1
python-pptx>=0.6.21
scikit-learn>=1.5.0
